LiveTV, unable to set a scheduled recording
LiveTV, button always showed play, not Schedule Recording/Cancel etc,
this caused it to stay at the retrieving for anything but a live
broadcast.

I had to go to 1.17.6 to set recordings, so this gives the basic
functionality back

TODO:
1. Refresh TV Guide to show new recording icons after schedule recording
selected or deleted.  Currently you have to go to the Home screen and
back in.
2 Show the current recording in the TV Guide as not included in the
.TimerId list once the recording starts.