Update VideoPlayer.brs
This change hides the screens that popup between items in a video playlist. The springboard, and the buffering progress bar.  If you would prefer not to hide the progress bar between list items, then I have another option available that will only hide the Springboard.  This option would only be a few lines of code.

I had to use roImageCanvas to be able to hide the buffering progress.  The user interactions I have put into place for this are somewhat crude, but any changes there would not be much work for me, or whoever else wanted to tweak it.