Fix Crashing of LiveTV
The problem is item.PlayOptions may be invalid

m.PlayOptions = Item.PlayOptions

The top piece of code CANNOT be used if item.playoptions is invalid. You need to check for this first, then if it isnt invalid allow the rest of the code to run. This corrects the problem and liveTV no longer crashes upon playing.